+++ logic/Preparation/Interface/ICharacter.cs
@@ -7,8 +7,8 @@ namespace Preparation.Interface
     {
         public long TeamID { get; }
         public int HP { get; set; }
-        public int Score { get; }
-        public void AddScore(int add);
+        public long Score { get; }
+        public void AddScore(long add);
         public double Vampire { get; }
         public PlayerStateType PlayerState { get; }
         public BulletType BulletOfPlayer { get; set; }
